bot_nonebot/requirements.txt
nonebot2[fastapi]>=2.4,<3.0
nonebot-adapter-onebot>=2.4,<3.0
httpx==0.28.1
